+++ x/slashing/test_common.go
@@ -94,9 +94,10 @@ func testAddr(addr string) sdk.AccAddress {
 
 func newTestMsgCreateValidator(address sdk.AccAddress, pubKey crypto.PubKey, amt sdk.Int) stake.MsgCreateValidator {
 	return stake.MsgCreateValidator{
-		Description:    stake.Description{},
-		ValidatorAddr:  address,
-		PubKey:         pubKey,
-		SelfDelegation: sdk.Coin{"steak", amt},
+		Description:   stake.Description{},
+		DelegatorAddr: address,
+		ValidatorAddr: address,
+		PubKey:        pubKey,
+		Delegation:    sdk.Coin{"steak", amt},
 	}
 }